+++ tests/conftest.py
@@ -3,10 +3,12 @@ import os
 import pytest
 
 os.environ["ROUTE_PROVIDER"] = "mock"
+os.environ["CALCULATION_MODEL"] = "v2.1"
 os.environ["GRAPHHOPPER_API_KEY"] = ""
 
 
 @pytest.fixture(autouse=True)
 def default_mock_route_provider(settings):
     settings.ROUTE_PROVIDER = "mock"
+    settings.CALCULATION_MODEL = "v2.1"
     settings.GRAPHHOPPER_API_KEY = ""